move lock functions into util.lock
